change packacking to war change settings to work as webapp add a test jsp file, and test static file add more pom depedencies restructure folders/rename files to controller/service/repo etc
