+++ src/main/java/com/rendezvous/controller/homeController.java
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.rendezvous.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+/**
+ *
+ * @author Leyteris
+ */
+@Controller
+public class homeController {
+
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+}
