+++ src/main/java/com/rendezvous/security/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.rendezvous;
+package com.rendezvous.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
