+++ src/main/resources/application.properties
@@ -1,9 +1,12 @@
 ## Spring DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
-spring.datasource.url = jdbc:mysql://localhost:3306/rendezvous?serverTimezone=UTC
+spring.datasource.url = jdbc:mysql://localhost:3306/stef?serverTimezone=Europe/Athens
 spring.datasource.username = root
-spring.datasource.password = ENAdioTRIA4!
+spring.datasource.password = 1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+
+spring.mvc.view.prefix=/WEB-INF/
+spring.mvc.view.suffix=.jsp
