+++ src/main/resources/static/newcss.css
@@ -1,0 +1,13 @@
+/*
+To change this license header, choose License Headers in Project Properties.
+To change this template file, choose Tools | Templates
+and open the template in the editor.
+*/
+/* 
+    Created on : Mar 7, 2021, 12:43:57 AM
+    Author     : Leyteris
+*/
+
+main {
+    color:black
+}
