+++ src/main/webapp/WEB-INF/home.jsp
@@ -1,0 +1,17 @@
+<%-- 
+    Document   : newjsp
+    Created on : Mar 7, 2021, 12:42:47 AM
+    Author     : Leyteris
+--%>
+
+<%@page contentType="text/html" pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <title>JSP Page</title>
+    </head>
+    <body>
+        <h1>Hello World!</h1>
+    </body>
+</html>
